Replace FILO linked list with FIFO ringbuffer
A linked list is (without additional measures) a first in last out data structure. That's not what we want for buffering commands. Therefor the linked list is replaced by a ringbuffer which is the most efficiënt first in first out data structure.

This ringbuffer is also faster than the previous linked list and prevents unnecessary memory fragmentation.

The only downside is that we need to claim (129 * buffersize) bytes of memory at front.

---

This PR is the second PR of a series working towards implementing rules.