Replace wifiManager with lean AP logic
The wifiManager is handy when that's all the logic
necessary for handling wifi.

However, the HeishaMon already has a pretty
extensive webserver and webpage, including
settings handling and such. When the rules
functionality has been added, having wifi is not
operation critical anymore. However, making sure
the HeishaMon doesn't reboot and block is.

The HeishaMon now just starts an AP when no
WiFi settings have been set. When they are set
the AP is only restarted after a certain period of
time, until it can connect to the external AP
again. Just as it currently works.

WiFi connection now can be configured in the
HeishaMon settings page, replacing the need for
the captive portal.

Not having to use the wifiManager also makes wifi
configuration fully not-blocking.