Fix related to WebSocket
I'm connecting to HeishaMon via an SSH tunnel. I access the application on a port other than 80 — in my case, it's 81. Unfortunately, the software doesn't work correctly and throws the following error:
Uncaught SyntaxError: Failed to construct 'WebSocket': The URL 'ws://127.0.0.2:81:80/ws' is invalid.
In the htmlcode.h file, the WebSocket connection unnecessarily forces port 80. This is not needed, because the port is automatically added in location.host as long as it differs from the default one (80 for ws and 443 for wss).